services/scraper/requirements.txt
scrapegraph-py>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
httpx>=0.26.0
redis>=5.0.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
celery>=5.3.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
pandas>=2.1.0
openpyxl>=3.1.0
